xfce4-settings: move to schnitzeltony's git repo to get new datetime setter
I implemented a datetime setting dialog based on gnome-control-center and
submitted a pull-request in [1,2]. As long as it is not submitted - or as long
as I find no way to patch binary files (e.g. icons - I asked for that in [3] -
no response yet) we use the git repo and a copied branch of the pull-request
(just in case some feedback from xfce-devels requires modifcations).

Some patches had to be modified/removed to apply against mainline HEAD still.
Furthermore in 0002-mouse-settings-dialog-add-touchscreen-pointer-option.patch
label 'alignment8' was introduced mainline and we do not need mouse-dialog_ui.h
when building from git in maintainer-mode - this file is created during build
process.

[1] http://mail.xfce.org/pipermail/xfce4-dev/2013-September/030442.html
[2] https://bugzilla.xfce.org/show_bug.cgi?id=10354
[3] http://lists.openembedded.org/pipermail/openembedded-core/2013-September/083804.html

Signed-off-by: Andreas Müller <schnitzeltony@googlemail.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>

Conflicts:
	meta-xfce/recipes-xfce/xfce4-settings/xfce4-settings_git.bb